+++ nucleus/src/arch/aarch64/memory/addr/virt_addr.rs
@@ -113,7 +113,7 @@ impl VirtAddr {
     /// See the `align_up` free function for more information.
     pub fn aligned_up<U>(self, align: U) -> Self
     where
-        U: Into<u64>,
+        U: Into<usize>,
     {
         VirtAddr(align_up(self.0, align.into()))
     }
@@ -123,7 +123,7 @@ impl VirtAddr {
     /// See the `align_down` free function for more information.
     pub fn aligned_down<U>(self, align: U) -> Self
     where
-        U: Into<u64>,
+        U: Into<usize>,
     {
         VirtAddr(align_down(self.0, align.into()))
     }
@@ -131,7 +131,7 @@ impl VirtAddr {
     /// Checks whether the virtual address has the demanded alignment.
     pub fn is_aligned<U>(self, align: U) -> bool
     where
-        U: Into<u64>,
+        U: Into<usize>,
     {
         self.aligned_down(align) == self
     }
